quiz_generator/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
fpdf
langchain-groq
langchain-core
python-multipart